Updated the monitoring window and added better graph support to imgui

+++ Engine/include/ECS/ISystem.h
@@ -5,5 +5,5 @@ class ECSRegistry;
 class ISystem {
 public:
     virtual ~ISystem() = default;
-    virtual void Update(ECSRegistry& registry, float deltaTime) = 0;
+    virtual void Update(ECSRegistry& registry, double deltaTime) = 0;
 };
