+++ Engine/include/Monitoring/Data/FrameStats.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct FrameStats {
-    float delta = 0.0f;
-    float fps = 0.0f;
-    float avgFps = 0.0f;
+    double delta = 0.0;
+    double fps = 0.0;
+    double avgFps = 0.0;
 };
